+++ lib/convo.js
