+++ lib/trends.js
