+++ lib/utility/random.js
