+++ lib/utility/timestamp.js
