+++ routes/timeline.js
@@ -3,9 +3,9 @@ var router = express.Router();
 var jSonParser = require('body-parser').json();
 var _ = require('underscore');
 // custom modules
-var users = require('../data/users');
-var parse = require('../data/utility/parse');
-var timestamp = require('../data/utility/timestamp');
+var users = require('../lib/users');
+var parse = require('../lib/utility/parse');
+var timestamp = require('../lib/utility/timestamp');
 
 // return an array of tweets
 function timeline(user, me, type) {
